Delayed movement start ⚠️  DO NOT MERGE!
If the pointer is far enough away, the cat will wait for a short time
before starting to move.

This was thought to maybe be nicer than the current behavior, as it
gives a short window of opportunity to place the cursor before the cat
starts going for it.

I find it to be really unintuitive and a step too indirect.
Now it's almost impossible to quickly react, should the cat get into
danger.
Therefore I'd leave it as-is, just keeping this commit for documentation
purposes.